rename block_elements -> elems
