+++ src/elems/select/multi/external.rs
@@ -74,7 +74,7 @@ impl<'a> External<'a> {
   ///
   /// # Example
   /// ```
-  /// use slack_blocks::block_elements::select;
+  /// use slack_blocks::elems::select;
   ///
   /// let select = select::External::from_placeholder_and_action_id(
   ///         r#"Hey I really would appreciate it if you chose
